+++ CHANGELOG.md
@@ -45,6 +45,7 @@ lists they are listed as multiple entries.
 - Added info modal to the playlist
 - Added initial partition support which allows you to connect to partition specified as a CLI argument
 - Added `listpartitions` CLI
+- Added Start and End Boundaries to ProgressBar increasing its Customizability
 
 ### Changed